+++ tests/BasicAuthenticationTest.php
@@ -451,7 +451,7 @@ class HttpBasicAuthenticationTest extends \PHPUnit_Framework_TestCase
                 "root" => "t00r",
                 "user" => "passw0rd"
             ],
-            "before" => function ($request, $response, $arguments) {
+            "before" => function ($request, $arguments) {
                 return $request->withAttribute("user", $arguments["user"]);
             }
         ]);