+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file, in reverse 
 - The `passthrough` setting was renamed to `ignore`.
 - Public setter methods `addRule()` and `withRules()` are now immutable.
 - PSR-7 double pass is now supported via [tuupola/callable-handler](https://github.com/tuupola/callable-handler) library.
+- PHP 7.1 is now minimal requirement.
 
 ### Added
 - Support for the [approved version of PSR-15](https://github.com/php-fig/http-server-middleware).
@@ -17,6 +18,5 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 - Most setters and getters for settings. Pass settings in an array only during initialization.
-- Support for PHP 5.X. PSR-15 is now PHP 7.x only.